New removeLiteral method in formula and made changes to setVar in dp
The removeLiteral method loops though the clauses and removes the
literal that matches the var in dpsolver setVar method. It uses the
evaluateClause method in clause to remove the literal

The setVar method checks if the passed in boolean value is true or
false. True sets the var (current level in tree) with a new true literal
otherwise false and removes the current var from the clauses that
contain the var.